Added doxygen comments to just about everything. Rewrote Fletcher8Source to use a sliding buffer.
git-svn-id: svn://graphite/concha/trunk/Concha@15 656cdb67-cf5e-0410-80f1-8fdcbb4d2177

+++ Fletcher8Source.h
@@ -18,16 +18,51 @@ typedef struct Fletcher8Source {
     Source source;
 	Source * from;
     int eof;
-    uint8_t x;
-    uint8_t y;
-    uint8_t z;
+    int x;
+    int y;
+    int z;
 	uint8_t a;
 	uint8_t b;
 } Fletcher8Source;
 
+/**
+ * Open an Fletcher 8-bit Source. The Source operates against an
+ * underlying Source while computing a Fletcher 8-bit checksum against
+ * the stream of incoming data octets.
+ * @param that points to the Fletcher 8-bit Source.
+ * @param from points to the underling Source.
+ * @return a pointer to the Descriptor Source as a Source.
+ */
 extern Source * openFletcher8Source(Fletcher8Source * that, Source * from);
+
+/**
+ * Read an octet of data from the Sink. The running value of the
+ * Fletcher 8-bit checksum is computed from the octet of data that
+ * was read from the underlying Source.
+ * @param that points to the Source.
+ * @return data as an integer for success, <0 otherwise.
+ */
 extern int readFletcher8Source(Source * that);
+
+/**
+ * Push an octet of data back into the Source. The octet of data is
+ * decomputed from the Fletcher 8-bit checksum.
+ * @param that points to the Source.
+ * @param data is an octet to push back to the Source.
+ * @return data as an integer for success, <0 otherwise.
+ */
 extern int pushFletcher8Source(Source * that, char data);
+
+/**
+ * Close the Sink. The final value of the Fletcher 8-bit checksum is
+ * computed. The a and b bytes of the checksum are written to the to
+ * Sink. The to Sink is closed.
+ * @param that points to the Sink.
+ * @return 0 for success, <0 otherwise.
+ * Close the Source. The from Source is closed.
+ * @param that points to the Source.
+ * @return 0 for success, <0 otherwise.
+ */
 extern int closeFletcher8Source(Source * that);
 
 #endif
